requirements.txt
flask
requests
reportlab
python-docx
pytesseract
pdf2image
Pillow
